+++ ios/Classes/PluginMetadata.m
@@ -7,7 +7,7 @@
     self = [super init];
     if (self) {
         _pluginPlatform = @"flutter";
-        _pluginVersion = @"7.2.0";
+        _pluginVersion = @"8.0.0";
     }
     return self;
 }