+++ lib/src/onfido.dart
@@ -7,8 +7,8 @@ class Onfido {
   /// - The [iOSLocalizationFileName] is the file name for configuring the localisation for iOS only. See [Language Customisation](https://github.com/onfido/flutter-sdk/#language-customisation) for the details.
   /// - The [iOSAppearance] is the configuration for iOS appearance, for Android you can see [UI Customisation](https://github.com/onfido/flutter-sdk/#ui-customisation) for the details.
   /// - The [mediaCallback] allows you to receive captured results and handle them in your application.
+  /// - The [biometricTokenCallback] allows you to manage the biometric token lifecycle and override the default implementation to read and write the token
   /// - The [enterpriseFeatures] is a set of features for advanced customisation, get in touch to enable this for your account.
-  /// - The [disableNFC] is a flag used to disable NFC (now enabled by default). Deprecated. Use `nfcOption` parameter to manage NFC settings
   /// - The [nfcOption] is the configuration for NFC (DISABLED, OPTIONAL, REQUIRED). You can see [Parameter Details](https://documentation.onfido.com/sdk/flutter/#parameter-details) for the details
   /// - The [onfidoTheme] is an enum used to determine which theme the UI will be rendered with (DARK, LIGHT, AUTOMATIC)
   Onfido(
@@ -16,26 +16,26 @@ class Onfido {
       String? iosLocalizationFileName,
       IOSAppearance? iosAppearance,
       EnterpriseFeatures? enterpriseFeatures,
-      @Deprecated("Deprecated. Use `nfcOption` parameter to manage NFC settings") bool? disableNFC,
       NFCOptions? nfcOption,
       OnfidoMediaCallback? mediaCallback,
+      BiometricTokenCallback? biometricTokenCallback,
       OnfidoTheme? onfidoTheme})
       : _sdkToken = sdkToken,
         _iOSLocalizationFileName = iosLocalizationFileName,
         _enterpriseFeatures = enterpriseFeatures,
         _iosAppearance = iosAppearance,
-        _disableNFC = disableNFC,
         _nfcOption = nfcOption,
         _mediaCallback = mediaCallback,
+        _biometricTokenCallback = biometricTokenCallback,
         _onfidoTheme = onfidoTheme;
 
   final String _sdkToken;
   final String? _iOSLocalizationFileName;
   final EnterpriseFeatures? _enterpriseFeatures;
   final IOSAppearance? _iosAppearance;
-  final bool? _disableNFC;
   final NFCOptions? _nfcOption;
   final OnfidoMediaCallback? _mediaCallback;
+  final BiometricTokenCallback? _biometricTokenCallback;
   final OnfidoTheme? _onfidoTheme;
 
   /// Start Onfido SDK using [FlowSteps].
@@ -47,7 +47,6 @@ class Onfido {
         iosAppearance: _iosAppearance,
         iosLocalizationFileName: _iOSLocalizationFileName,
         enterpriseFeatures: _enterpriseFeatures,
-        disableNFC: _disableNFC,
         nfcOption: _nfcOption,
         mediaCallback: _mediaCallback,
         onfidoTheme: _onfidoTheme);
@@ -60,6 +59,7 @@ class Onfido {
         workflowRunId: workflowRunId,
         iosAppearance: _iosAppearance,
         mediaCallback: _mediaCallback,
+        biometricTokenCallback: _biometricTokenCallback,
         iosLocalizationFileName: _iOSLocalizationFileName,
         enterpriseFeatures: _enterpriseFeatures,
         onfidoTheme: _onfidoTheme);