+++ lib/src/serializer/start_studio_serializer.dart
@@ -1,6 +1,6 @@
 import 'package:onfido_sdk/onfido_sdk.dart';
 
-/// convert dart code to dictionary to be passed to the paltform
+/// convert dart code to dictionary to be passed to the platform
 class StartStudioSerializer {
   static serialize(
       {required String sdkToken,
@@ -8,6 +8,7 @@ class StartStudioSerializer {
       IOSAppearance? iosAppearance,
       String? iosLocalizationFileName,
       bool? shouldUseMediaCallback,
+      bool? shouldUseBiometricTokenCallback,
       EnterpriseFeatures? enterpriseFeatures,
       OnfidoTheme? onfidoTheme}) {
     return {
@@ -16,6 +17,7 @@ class StartStudioSerializer {
       'iosAppearance': iosAppearance?.toJson(),
       'iosLocalizationFileName': iosLocalizationFileName,
       'shouldUseMediaCallback': shouldUseMediaCallback,
+      'shouldUseBiometricTokenCallback': shouldUseBiometricTokenCallback,
       'enterpriseFeatures': enterpriseFeatures?.toJson(),
       'onfidoTheme': onfidoTheme?.name
     };