+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [8.0.0] - 2024-11-7
+
+- Removed `disableNFC` parameter (Use `nfcOption` with `NFCOptions.Disabled` value instead)
+- Update underlying Onfido native SDK versions:
+  - Android 22.0.+ (up from 21.2.+)
+  - iOS 32.0.+ (up from 30.6.+)
+
 ## [7.2.0] - 2024-09-09
 
 - Update underlying Onfido native SDK versions: